\name{RcppRP-package}
\alias{RcppRP-package}
\alias{RcppRP}
\docType{package}
\title{
Random portfolios using Rcpp
}
\description{
This package provides functions to generate random portfolios and calcuate a constrained objective function over the set of random portfolios.
}
\details{
\tabular{ll}{
Package: \tab RcppRP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-09\cr
License: \tab What Licence is it under ?\cr
}
A function to generate random_portfolios already exists in PortfolioAnalytics, and the intent of this package is to implement the algorithm in C++ to improve performance.
}
\author{
Ross Bennett

Maintainer: Ross Bennett <rossbennett34@gmail.com>
}
\references{
PortfolioAnalytics
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in
%% ~~ the R documentation directory ~~
\keyword{ package }
\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
