\name{rp_optimize_v1}
\alias{rp_optimize_v1}
\title{V1 Random portfolio optimization}
\usage{
  rp_optimize_v1(R, objective, search_size = 2000,
    portfolio = NULL, rp = NULL, momentFUN = "set_moments",
    ...)
}
\arguments{
  \item{R}{}

  \item{portfolio}{}

  \item{search_size}{}

  \item{objective}{}

  \item{rp}{}

  \item{\dots}{}
}
\description{
  TODO Random portfolio optimization description
}

