\name{rp_optimize_v2}
\alias{rp_optimize_v2}
\title{V2 Random portfolio optimization}
\usage{
  rp_optimize_v2(R, objective = NULL, portfolio = NULL,
    search_size = 2000, rp = NULL,
    momentFUN = "set_moments", ...)
}
\arguments{
  \item{R}{}

  \item{portfolio}{}

  \item{search_size}{}

  \item{objective}{}

  \item{rp}{}

  \item{\dots}{}
}
\description{
  TODO Random portfolio optimization description
}

